Move Parser::siimple_command
